requirements.txt
streamlit
llama-index
llama-index-vector-stores-chroma
openai
chromadb
python-dotenv